REF: Move conversion functions to core/convert.py
Move conversion function to core/convert.py
Restore docs for convert_objects
Add bug fix notes for restores convert_objects
Ensure copy is passed to _possibly_convert_objects when needed
